requirements.txt
# F1 2025 Telemetry Dashboard
websockets>=14.0
pyyaml>=6.0
azure-ai-projects>=2.0.0
azure-identity>=1.15.0
